requirements.txt
transformers
torch
chromadb
sentence-transformers
fastapi
streamlit
streamlit-monaco
pytest
pyyaml
accelerate
bitsandbytes
google-genai
pyngrok
